requirements.txt
aiofiles==23.2.1
beautifulsoup4==4.12.3
duckduckgo-search==4.4.3
faiss-cpu==1.8.0
fastapi==0.110.2
httpx==0.28.1
langgraph==0.2.15
numpy==1.26.4
lxml==5.3.0
lxml-html-clean==0.4.3
playwright==1.43.0
pydantic==2.12.0
pydantic-settings==2.11.0
python-dotenv==1.0.1
readability-lxml==0.8.1
requests==2.31.0
rich==13.7.1
sentence-transformers==2.5.1
tenacity==8.2.3
tqdm==4.66.2
typer==0.12.3
uvicorn[standard]==0.37.0
pytest==8.1.1
pytest-asyncio==0.23.5
openai==1.107.3
openai-agents==0.3.3
